Added option to choose protocol version in cf-net
Added option to choose the maximum CFEngine protocol version to use in
cf-net. Maximum, because the server can negotiate on an older version.

Currently, there is no way to test older protocol versions without
targeting an older server. This option will make that possible.
Furthermore, it can be used as a workaround in the potential cases where
we issue a new protocol that does not work.

Ticket: ENT-12519
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>