Now `cf-net get` no longer unlinks original file
The command `cf-net get <FILENAME>` now writes to `<FILENAME>.cfnew`
followed by replacing `<FILENAME>` with `<FILENAME>.cfnew`. Previously
it would unlink and re-create the original file.

With the new `"filestream"` protocol, we need the original file in order
to compute its signature, so that we can save bandwidth using the RSYNC
protocol. Furthermore, unlinking the original file, causes a brief
moment when it either does not exist or is incomplete. This can be
problematic if other processes depend on it.

Ticket: ENT-12511
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>