Link cf-agent and cf-serverd with leech if available
In order to use leech in the enterprise plugin, we need to link with
leech in core. I added code in configure.ac to link with leech if it is
available. A community version of CFEngine will work just fine without
it. However, an enterprise version will require it.

Ticket: ENT-11574
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>