Fixed bug causing internal server error on remote file copy
Previously I had a check to see if the input buffer is full. This was to
detect infinite loops where the job needs more data, but we fail to
provide it. However, it appears that even though the input buffer is
full, the job is still able to complete if you just give it a few more
iterations. Futhermore, librsync has an internal detection system that
returns `RS_INTERNAL_ERROR` if the job cannot make progress. Thus, we
don't have to worry about infinite loops.

Ticket: CFE-4507
Changelog: Fixed bug where remotly copying large files consisting of
           random bytes would cause internal server error.
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>