Fixed invalid colon separating host and path from cifs and panfs in fstab
This commit introduces a conditional check to format the device string for CIFS and PanFS without a colon (:), as required by their specific syntax. It also updates the logging to show the full fstab entry that is being added.

Additionally, the log message has been updated to be more descriptive, now displaying the full fstab entry that is being added.

Ticket: CFE-4494