Fixed bug where rename fails to reset temporarily cleared immutable bit
The problem was that it tried to reset the immutable bit status of the
old file, instead of the new file after calling rename(). The old file
obviously is not immutable because it is the temporary file that we want
to replace the original file with after e.g., rendering its contents.

Ticket: ENT-13179
Changelog: Title
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>