Fixed cf-support usage of coredumpctl matching
Initially we had coredumpctl info /var/cfengine/bin/cf-* which matched the EXE (executable name).
We then changed this to /var/cfengine/* which was an incorrect change which limited the EXE to only items at /var/cfengine/ since the star is a glob and expanded to the files in that directory.

(commit: 68b8429a85a86f88592a96f5ff3a0f6b857465c9)

Instead we should use a find command to find files that are executable under /var/cfengine and run coredumpctl on each one we find.

Ticket: ENT-13272
Changelog: title