Adjusted selinux policy to be more resilient to updates
In rhel-10 the current policy errored out due to missing key_socket class.
Because we were specifying each class explicitly we were prone to trouble when changes happened in the kernel and selinux-policy (the package/repo which builds the .pp policy file).
Replacing all of the class elements with a single all_kernel_class_perms macro will allow us to get what we want: all of the classes available for use in our policy but in a way that will change outside of our policy.

Ticket: ENT-13016
Changelog: none