Fixed length checking in StatFile
If a symlink resolves to a path that is longer than the maximum allowed by the protocol, currently 4088 bytes, then an error response is generated.

Previously the server side would ignore an error from SendTransaction() due to the large size and the client-side would hang/timeout waiting for a response.

Ticket: ENT-13542
Changelog: title